requirements.txt
psycopg2-binary==2.9.10
pytest==7.2.2
python-dotenv==1.0.0
tabulate==0.9.0

